dedede

+++ index.js
@@ -328,16 +328,17 @@ app.post('/api/sendmetafield', function(req, res) {
           .then((response) => {
             res.json(response.data);
 
-            cloudinary.api.delete_resources_by_prefix(`customers/${customerId}/${oldGtkey.split(":")[1]}`)
-              .then(() => {
-                console.log(`delete - customers/${customerId}/${oldGtkey.split(":")[1]}`)
-                cloudinary.api.delete_folder(`customers/${customerId}/${oldGtkey.split(":")[1]}`)
-                  .then((result) => {
-                    console.log(`delete - customers/${customerId}/${oldGtkey.split(":")[1]}`)
-                    res.json(result);
-                  });
-              })
-
+            if (customerId && oldGtkey.split(":")[1]) {
+              cloudinary.api.delete_resources_by_prefix(`customers/${customerId}/${oldGtkey.split(":")[1]}`)
+                .then(() => {
+                  console.log(`delete - customers/${customerId}/${oldGtkey.split(":")[1]}`)
+                  cloudinary.api.delete_folder(`customers/${customerId}/${oldGtkey.split(":")[1]}`)
+                    .then((result) => {
+                      console.log(`delete - customers/${customerId}/${oldGtkey.split(":")[1]}`)
+                      res.json(result);
+                    });
+                })
+            }
           });
     });
   }
